fix main stats components
