--- client/src/app/stats/StatsCard.tsx
@@ -1,23 +1,0 @@
-"use client";
-
-import { Card, CardContent, CardDescription, CardHeader, CardTitle } from "@/components/ui/card";
-
-interface StatsCardProps {
-  title: string;
-  desc: string;
-  value: string | number | null;
-}
-
-export const StatsCard = ({ title, desc, value }: StatsCardProps) => {
-  return (
-    <Card>
-      <CardHeader className="pb-4">
-        <CardTitle>{title}</CardTitle>
-        <CardDescription>{desc}</CardDescription>
-      </CardHeader>
-      <CardContent>
-        <div className="text-4xl font-bold">{value}</div>
-      </CardContent>
-    </Card>
-  );
-};
